requirements.txt

streamlit==1.29.0
pandas==2.2.2
numpy==1.26.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.103.0
bcrypt==4.0.1
cryptography>=41.0.0,<43.0.0
flask==2.3.3
python-dateutil==2.8.2
schedule==1.2.0
psutil==5.9.6
apscheduler==3.10.4
bleach==6.1.0
flask-cors==4.0.0
fpdf2==2.8.3
passlib==1.7.4
plotly==5.17.0
pyjwt==2.8.0
streamlit-aggrid==0.3.4
#fpdf
fpdf
python-multipart
python-multipart
uvicorn
bcrypt
fastapi
passlib
pydantic
python-jose
python-multipart
sqlalchemy
uvicorn
fastapi
httpx
pydantic-settings
pytest
pytest-asyncio
pytest-cov
sqlalchemy
uvicorn